backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.46
pymysql==1.1.2
cryptography==46.0.3
passlib==1.7.4
bcrypt==5.0.0
pyjwt==2.10.1
pydantic==2.12.5
email-validator==2.3.0
python-dotenv==1.2.1
